requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
supabase==2.3.4
google-generativeai==0.5.2
